+++ llumnix/entrypoints/vllm/api_server.py
@@ -27,13 +27,12 @@ from vllm.sampling_params import SamplingParams
 from vllm.engine.arg_utils import AsyncEngineArgs
 from vllm.engine.async_llm_engine import AsyncStream
 
-from llumnix.arg_utils import EngineManagerArgs
+from llumnix.arg_utils import LlumnixArgumentParser, EngineManagerArgs
 from llumnix.entrypoints.llumnix_entrypoints import (get_ip_address,
                                                      launch_ray_cluster,
                                                      connect_to_ray_cluster,
                                                      is_gpu_available,
                                                      init_llumnix_components,
-                                                     LlumnixArgumentParser,
                                                      init_per_token_latency_breakdown_dict,
                                                      record_per_token_latency_breakdown)
 from llumnix.logger import init_logger
@@ -278,7 +277,6 @@ if __name__ == "__main__":
     if cfg.RAY.LAUNCH_RAY_CLUSTER:
         # Launch the ray cluster for multi-node serving.
         launch_ray_cluster(cfg.RAY.RAY_CLUSTER_PORT)
-
     # Connect to a ray cluster.
     connect_to_ray_cluster(port=cfg.RAY.RAY_CLUSTER_PORT)
